Moved HistoricalDataService and updated readme
